// src/api/models/todo-body-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * S2I final
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TodoBodyReq
 */
export interface TodoBodyReq {
    /**
     * 
     * @type {string}
     * @memberof TodoBodyReq
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof TodoBodyReq
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof TodoBodyReq
     */
    'note'?: string;
    /**
     * 
     * @type {string}
     * @memberof TodoBodyReq
     */
    'category'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TodoBodyReq
     */
    'checked'?: boolean;
}